"12:40pm. 51m in. This proof that circle is equivalent to integer is complicated. I can't really follow it right now and I really want to play with this in a proof assistant instead to make sure of what the rules are. But finally here is the univalence axiom.
This must be like the first time it was actually used in the entire course.

1:10pm. Fuck, done, done, I am done!

Finally! As expected even as programming, math is not to my taste. And yet I have the misfortune of having to go in this direction.

I might have a purpose in life, but I wish my brain was structured better for the demands of the task. But there is no need to feel adrift about this. I just have to do it.

Well, it will get easier as I get some real skills in using these new systems. For now, I will just have to bear with it.

Today (and yesterday) I watched the 3 lectures in absolute record time so I have the option of calling it a day early, but 1:15pm is just too soon no matter how one looks.

Even so I just can't keep my foot on the pedal all the time, so I am going to take a break here for an hour or two. When I resume I am going to watch the two talks.

https://www.youtube.com/watch?v=WfDcrN5_1wA
What do we mean by "equal" - Pierre Deligne

https://www.youtube.com/watch?v=3iKwq9iRHWk
Carlo Angiuli, Computational semantics of Cartesian cubical type theory

This will be my last bit for the day. Tomorrow I will go through the OPLSS 18' lectures by Harper.

And then...I can start writing the monthly review a few days early. Then it is programming time. I am going to beat software foundations. And I will put the past where I could not do it to the rest."