"1:45pm. https://youtu.be/RhDuRmg-SdA?t=1329
Here he talks about entailment as a ninja move. "Try to explain it to your colleagues - you are not going to get anywhere."

To be honest, I am not sure that I myself understand entailment. I have an idea that sort of makes sense, but I can't say that I am certain about what it should be. I've seen plenty of examples of it by now though.

https://youtu.be/RhDuRmg-SdA?t=1352

Well, yes, I was thinking it as maps in part...

2:20pm. https://youtu.be/RhDuRmg-SdA?t=2925

I am lost as to what this part here is.

2:25pm. Ah, let me take a little break here. Only 35m left so I might as well do it now. I did not have the chance to have one earlier.

3:15pm. https://youtu.be/RhDuRmg-SdA?t=4911

Nearly done.

"...And then you have full higher dimensional type structure that gives computational meaning to HoTT by distinguishing paths from identities and getting that all sorted out so that J works the way it works, but that has nothing to do with J - don't worry about it. We instead have path structure. And in fact, what happens is that the HoTT formalism becomes an uncomfortable peephole that only lets you look at points. You can never ever look at lines directly - you can only look at them once they have been quoted into a path type. And that's clumsy as hell. You don't want to work like that. So the current thinking is, what you wanna do is work in the cubical type theory.

And you work with the cubes directly. Forget the old formalism. Those are done. That's our current operating assumption."

This last bit is some good information as to where to proceed with my study once the time comes. I'll forget the HoTT stuff as suggested and try to learn CTT.

https://youtu.be/RhDuRmg-SdA?t=4996

"So that is the state of the play in like, very, absurdly advanced type theory in some respects. If you ask me what is it all good for, well, I do not have a good answer for you. Someone was asking me this. And the only thing, I can use a cheap trick which was you know...in 1980 when I first heard about type theory and first heard that proofs should be programs, 'goddamit', and the law of excluded middle is not valid - I was delighted because that's what I thought I just didn't know why. I hated proofs by contradiction. I thought they were the biggest cheat in the world and I just felt like it shouldn't be allowed. It's not fair. And then Constable told me you are right. And I was like - this is the best thing I've ever heard of."

He hopes it might turn out to be useful at some point, like 30-40 years time.

Oh boy, that is some time horizon. I take it he is not expecting the Singularity any time soon.

If the Singularity takes any time that long, I am going to be the biggest fool in the world, or maybe a stepping stone for the next seeker. Either way, if it cannot be me then I have to do as much as possible so that the next in line grasps it. If it will be me, then I have to do everything in my power to not miss the chance.

And I will know once I master the use of type theory whether I really have what it takes to do this.

3:35pm. Let me skim the talk by Angiuli and then I will be done for the day. Maybe I will watch it in full depending on how it goes."