"8:55am. Let me slowly start the review.
"During the first week of June I finally finished the first [Software Foundations](https://softwarefoundations.cis.upenn.edu/) volume. I started it in late April, so it took me about 1.5 months to really deal with it in full. My conclusion is that the IndProp chapter is significantly harder than all the other chapters in the book combined, so be warned if you are thinking of going through the book at some point. In comparison, the first [Agda SF](https://plfa.github.io/) volume only took me ten days.

It was damn tough to get through all of this and I am going to have to go through the rest of the volumes as practice. I am not that interested in the PL parts of it; my desire is to gain the skill to formalize arbitrary math that I run into in the wilds. Nonetheless, there are only so many source on dependently typed programming at this moment, so doing it like this is probably better than studying random libraries. I'd like to thank all those who have been answering my amateurish question on SO. I'd never get anywhere without you helping me get over the humps.

Given how hard SF (in Coq) was, for anybody interested in this it might be better to start with [Lean instead](http://leanprover.github.io/logic_and_proof/) as I've heard some good things about it. I have no experience with it yet personally though.

For the last 3 weeks I was in a study mood. I've been doing logic puzzles non-stop prior to that, so as a change of pace I wanted to study [homotopy type theory](https://www.reddit.com/r/ProgrammingLanguages/comments/c2vii4/univalent_foundations_new_foundations_of/) just to see what awaits me after I get through Software Foundations. The best primer at this point are Harper's [2013 HoTT lectures](https://www.cs.cmu.edu/~rwh/courses/hott/). With the advent of [cubical type theories](https://www.youtube.com/watch?v=RhDuRmg-SdA) that have a computational interpretation they are somewhat out of date, but they are enough to get a shaky grasp on the subject.

There is only so much one can learn by just talking about the subject - I really find it miserable when he does programming on the blackboard, so to get a real grasp I am going to have to try programming in a language with a cubical type system. At this point that means [Cubical Agda](https://github.com/agda/cubical), but I'll also check out the ones ([RedPRL](http://www.redprl.org/en/latest/), [RedTT](https://github.com/RedPRL/redtt)) by Harper's group.

My hunch is that univalent math will be necessary for what I want to do. Given that it allows for a novel definition of real numbers, I am guessing that this capability will also extend to dealing with other structures that have bounded infinities as a part of their composition such as probability distributions.

We'll see how it goes. I have a detailed plan in mind, but it all depends at whether I can attain the requisite level of skill in this subject to deal with the individual steps. I need a new kind of reasoning to be able to do this, so for the time being I will just be plugging away at it. It took me a long time to get good at regular programming, so I am not bothered that it will take me a while to get good at this as well.

Maybe it will come to nothing and I will turn out to be a madman just doing meaningless logic in the end. I cannot tell whether I am rising to the heavens or sinking to the bottom of hell. But as usual, the duty calls and the desire in me keeps pushing me forward to climb that mountain.

A language like [Spiral](https://github.com/mrakgr/The-Spiral-Language/tree/v0.1) I managed to do, but cracking credit assignment is such a tall order even for me."

10:15am. I thought it would be longer than this, but this is perfectly fine. I can talk about the details of my plan as I go along. I'll leave it for the following updates.

10:25am. I sort of did what I set out for myself over the day, but I still have a lot of time left over. Compared to the sweltering heat of the last few days, it is so nice and cool right now.

I can't just stop here for the day - the circumstances do not matter. I have to keep going forward. But rather than pluging right into the second SF volume, let me take a look at the Lean book. I'll start the second volume later in the day."