"11:05am. Let me have breakfast.
https://mangadex.org/title/22598/hanakaku-the-last-girl-standing

I am actually into this.

https://weightagnostic.github.io/

Oh, this is interesting. I had a similar idea, but discarded it as being too unlikely.

Edit: Academics really have an easy life to be able to play like this. It is interesting, but this is all it is. I cannot accept these sorts of methods. It reminds me of the old evolutionary papers before deep learning became popular.

It really all comes up to credit assignment in the end doesn't it? The is a remarkable lack of imagination how this might be done - the field as a whole that it sort of stumbled into backpropagation, but backpropagation is a fairly advanced thing. Not only does it require reals, but it is not really designed for optimizing static structures. It can act like one, but it is not a real credit assignment scheme.

12:15pm. I am really into Hanakaku. Let me continue reading this and then I will read the Lean book.

12:35pm. https://mangadex.org/chapter/210986/5

I really changed so much in the past 4.5 years. I really have come to like honest, evil girls like her. If this is her reason for being despicable then I am ok with it.

As I get older my taste gets more refined. There is nothing to gain from good people winning, so why not praise evil for a change?

1:10pm. https://mangadex.org/chapter/401684/1

Final chapter. I'll leave it for later.

Let me read the Lean book for a bit.

It is not just that the manga is good. It is just that when I am rested I find manga so much more enjoyable. When I am tired I really can't immerse myself properly. It is always like that. So doing this in the morning is definitely the best kind of reward for me.

Also I've decided. It must have been years ago, but I am going to pick up the old HoMM 3 campaign that I never finished. Doom has really gotten boring, and I am out of things to do past 6pm.

1:40pm. Let me finally start the working day. Today I've been chilling, but let me do a little reading of the Lean book. I am not interested in doing much with it at the moment, I'll just read a chapter or two to see what it is about. Then I'll do a few exercises from the SF vol 2 and that will be it for the day. Tomorrow I will start in full force."