"8:55am. As it turned out, my HoMM 3 save was nearly six years old. I really got into the game and ended up laying till past 12am. I really greatly enjoyed the session. Getting so much into a game is truly the best part of life. This right here is what I am fighting for.
This is the true me. The only thing I am missing is the exploratory element where as a kid I went from game to game. But I will get that back - once I crack credit assignment.

9:15am. Let me start the day here. I won't say that it is time to resume work. Rather it is time to resume training.

I am going to go through the first chapter today and then move to the Agda book."