"10:05am. I did not think I would get up so late today. I might as well have that breakfast now.
https://arxiv.org/abs/1906.08857
Deep Neuroevolution of Recurrent and Discrete World Models

Let me read this paper.

Maybe it would be worth considering evolution at some point, but really, credit assignment is needed to scale. This stuff is just filler science.

11:25am. Done with breakfast. Let me start. Last night I overdid and went to bed past 1am.

The Fall Of Sandro was such a disappointing way to end the campaign. He spent all his time behind a barrier and turns out to be a weakling. I actually won even though I attacked him by accident without the main army. Lame.

I have fond memories though of campaign beginnings. On the hardest difficulty there is a big mismatch in starting resources and if you just wait and build up slowly the computer will overwhelm you. I found this really paradoxical as a kid, but the way to win in such a situation is to be really aggressive. Cheese things with spells, blitzkrieg the mobs without losing any troops, take anything that is not bolted down and then it becomes possible to win.

Later on though, when the main heroes have >20 in every stat and high level spells, they start carrying the battles on their own and the game becomes trivial, but the moments of weakness are where the fun is.

11:50am. Well, it is time to resume my Coq adventure. Those advanced exercises are not going to do themselves.

Let me put in some time into it."