"6:35pm. Done with lunch...and the Symphoraw is out.
6:50pm. 50% done. Let me play Homm 3 in the meantime.

10:25pm. Done with today's session. I really wanted to beat that huge army on the first map of Sceptre of Power and I managed it - using the power of logistics. As expected Town Portal is extremely powerful on a strategic level. It allowed me not only to concentrate my own forces, but take over town in the enemy territory. For every enemy town town that means that 14k of gold that would have gone to the enemy every week goes to me instead.

I lead the op enemy hero by the nose until I amassed enough troops, and baited it to attack me in one of the castles. Then after that it was a matter of using berserk to take out its catapult and archers while the troops in the castle took care of the fliers. This allowed me to completely avoid fighting the footmen. Once the archers and the fliers were dead, the enemy had no choice to retreat and abandon the ground troops. This is not a small thing as together they had ~10k in hp. They would have definitely damaged me greatly had I slugged it out with them.

I finished the map with 3 heroes at level 10, all the powerups and spells from the guilds collected. Moreover, those 3 all have Town Portal. I am now in truly excellent shape as far as the following maps are concerned. Not to mention all the heroes have Earth Magic + Intelligence.

Last time I did this map - well, I had Armageddon + Antimagic, so it is not like I did too badly, but I had to rush the level. With a cheat combo it would have left me in a bad condition afterwards.

I really like this as this is really the first time I won through sheer resource allocation. I was truly a satisfying way to play this.

10:40pm. Often, how the rest of the campaign goes is really determined by the first map - the spells and the skills are carried from that one and they have a big influence on the hero's effective power. Beating the maps on Impossible really does require knowing how to use spells effectively.

10:50pm. https://mangadex.org/title/34522/himenospia

Oh, more of this came out. Actually I will leave it for tomorrow. I want to eat and go to bed.

7/7/2019

7:50am. I got up too early today. Let me chill for a while before I start. Yesterday I only watched the Symphogear raw, so let me watch the real thing now. For Himenospia, now that the translation is caught up, I guess like with Asuka the wait will have to start. My thanks goes to the hero for picking this up.

I really should not have watched the Symphogear raw, but I remember last time how long it took the subs to come out. Really, let me say something nice about Crunchyroll. It is too bad about it killing the subbing community, but CR is 1000x times better than those shit tier companies that impose years worth of delay on licensed manga and novels. CR is awright.

8am. Let me start the chilling. I'll finish the last two exercises after that and move on to the next chapter. It is a lot shorter than the previous ones, so I might be able to do it today assuming the exercises do not turn out too hard."