"8:45am. Let me finally watch that ep of Granbelm and then I will start. It is going to be a long month, so I need some fuel for this.
9:15pm. The series feels as if it will be mediocre.

Before I start, let me put in a few words about ML. I feel that the basic principles of improvement through gradual changes relative to some goal are grasped well enough, but the principles of hierarchy needed to make the whole thing efficient are a complete mystery.

Without those principle any gains the system makes are transient. Make no mistake, if there was real learning going on in a system, you would not be dumping it and starting from scratch on some new dataset. Real learning is accumulation and refinement without end.

If you had a system that could learn for real, you'd train it on both Atari and Go.

The end goal has always been omnipotence, but what I really, truly found difficult was to say a single non-trivial thing about a single step forward. Backprop is such a vast improvement over blind evolutionary mutation, but it is just a hack derived from calculus. It is just an embarrassment to be an intelligent being and have nothing at all to be able to say about intelligence itself.

I do not understand why this is so difficult? Why is the principle of hierarchy so elusive?

The principle of hierarchy is also the principle of communication, so that means even more that I cannot avoid tackling it. My entire life was a waste. I should have just charged straight through at this lance in hand. I should have anticipated that I would have to be doing this anyway.

The story that is Simulacrum was a milestone in my life. It is the first time I managed to envision what it would be like to be inside the algorithm. The Inspired are just such a definite step forward conceptually from all the other superpowered characters. They embody not the perfection, but what it really means to make a single step forward. Their concept was what convinced me that superpowers lie could lie not only in the domain of imagination, but in reality.

It is my great yearning and yet even though I can see so many things that lie beyond it, I cannot make that tiny step forward towards it. I put in so much effort just so I could get so much little. But that little thing is the foundation of everything.

9:40am. From my present perspective it is such a tragedy. I cannot see the point of me not being able to do even that much. I must seem like a complete fool to anybody simulating this universe from the start. I loathe my fate. My life is nothing, but hardship. The culture that I love is nothing, but weakness. I cannot accept it. And I cannot forgive this.

9:45am. So because of that, indirectly what I really need to make single step forward turns out to be math. The kind of programming that comes naturally to me is useless for what I want to do.

Logic will be the light shining down the road to power.

I really do feel bad about doing this so late. I am sorry for trying to avoid this and trying to take all sorts of shortcuts to avoid doing this. I want to send this apology to the world and my past self.

This heavy burden is my to bear. It is my responsibility that I do this. I am the one who has to do this and nobody else.

The reason why my life has been a failure is because I could not just make up my mind to tackle my math deficiency head on. All my past failures have been because I shirked from the duty of trying to illuminate intelligence which I knew would give me everything and instead decided to let the others waste their time on it. I should have known better. I should have known that this is the one place where I should have put my faith in.

It is all because of this that during my trading days instead of trying to automate my decisions, that I decided to approach the market as a gamer. **I trusted the machines, but I did not trust myself to use them well.** That is why I did not become a programmer. My talent really worked against me here. I reasoned that since I was intelligent, but my agents would not be that I should minimize my reliance on them.

This is the greatest mistake that I made in my life.

In the short term it is always easier to rely on one's own abilities rather than on machines. It is in the long term that the benefits of automation are accrued. The same story goes again for regular programming vs math.

10am. Real life is not a game - real life is about getting to the root, while this is trivially easy in games through cheats. My whole life up to now can neatly be interpreted as a cautionary tale against being bad at math.

I will accept this interpretation.

I need to always remind myself what I am doing this for. The light of understanding that I will attain through this hardship will allow me to win the battles that lie ahead. I might be a loser now, but once I get it, I will become invincible. Mastering type theory and after that, through it math is the final stretch. I will do this, find the principles and then aright my life on the tracks that should always have been there.

10:20am. This was a good rant. I am such a weakling, but if I can internalize what I wrote above there just might be some hope for me. The weight of my regrets is a heavy burden on me, but once I finally start behaving in a rational manner all of that will be left behind. I will become who I want and need to be.

Now it is time to start."