"8:55am.
```
Theorem hoare_asgn : forall Q X a,
  {{Q [X |-> a]}} X ::= a {{Q}}.
Proof.
  unfold hoare_triple.
  intros Q X a st st' HE HQ.
  inversion HE; subst.
  unfold assn_sub in HQ. assumption.  Qed.
```

I said I would try to get a computational perspective on this, but I really am confused as to what the effect of equality is. Can I really assume that it is doing nothing? Would everything here get compiled to the identity or would something more happen?

The only way I will be able to figure this out is to make my own dependently typed language.

I'd really like to avoid spending a few years on this like I did with Spiral.

I really just want the learning algorithm now. The me of the future will have all the time in the world to play with languages after cracking ML, so I'd really prefer if the universe did not burden me with this.

9:30am. https://www.reddit.com/r/MachineLearning/comments/cbgizh/d_controversial_theories_in_mlai/

My focus is low and I find myself reading this thread.

neural_kusp_machine: "Honestly, too many to list.

My favorite one has to be the two extreme views on machine learning. I strongly believe that ML, as a field, is essentially concerned with a statistical problem (sample complexity, statistical learning), and also a computational problem (efficiently finding a good model).

Each of these two problems, if viewed independently, is trivial. Optimal sample complexity can be achieved by searching over universal models of increasing complexity (e.g. running all possible computer programs with increasing length, until one fits your data). Optimal search is trivial if your model is a look-up table (it perfectly fits any training data by definition).

Some researchers strongly believe that we should only care about the statistical problem and that ML=Stats. Some very controversial arguments include: if your model class is universal 'enough', you can cook up a distribution over models -- biased towards simpler ones -- and randomly sample models, which will be good with non-zero probability. This means that instead of training networks, there might be some 'smart' way to randomly sample the parameters to get a good model in terms of training and test performance.

Others strongly believe that the computational problem is what matters, and that ML=CS (or, perhaps more commonly, ML=Optimization). The argument is that, if your search is efficient and biased towards simpler models, then you don't have to care about regularization, the model's capacity, overfitting and so on. The strongest evidence for this view is that huge neural networks don't overfit. There is a new wave of ideas that start by constructing a model that perfectly fits your data (which is now easily to do in closed-form with neural networks), and then optimize its 'smoothness' to make it generalize -- the argument, again, implies that you can solve ML with optimization.

It is very clear that in reality the ML=Stats people also deal with computation (how do you cook up that prior over models? how do you sample them? how many times do you sample? how do you check whether a model is sufficiently good?), while the ML=CS people also deal with stats (why does 'smoothing' out a network make it generalize? what does 'simple' mean? why does 'simple' mean good generalization?), but many researchers disagree, seeing the other problem as a 'small burden' in ML, and not as the core part.

Even more controversial is the view that ML is solved (up to constant factors). There are universal search algorithms (proposed by Hutter, Schmidhuber, and so on, based on Levin Search) which are sample-optimal and efficient for any problem, but can take exponentially longer to execute -- the catch is that they are exponentially slower in terms of the minimum description length of the best model, which is not captured by 'efficiency', as it is virtually independent of input size, dimension and so on. Very few people believe this nowadays, and most serious researchers ignore the existence of these universal learners -- or at least consider them a 'theoretical hack' that should be disregarded. However, it lead to many fruitful ideas that the main problem of ML is one of minimum description length -- once we have a language where good models for natural problems can be easily described, ML is solved. Although controversial, I strongly believe this might be the best way to describe the philosophical obstacles of learning."

It is interesting.

Let me finish the first Hoare chapter. I am going through it quickly enough.

9:40am. Focus me. Just do this.

9:55am. Done with the first chapter. I am not getting much out of this to be honest.

10:15am.

```
(** In general, we say that "[P] is the weakest precondition of
    command [c] for postcondition [Q]" if [{{P}} c {{Q}}] and if,
    whenever [P'] is an assertion such that [{{P'}} c {{Q}}], it is
    the case that [P' st] implies [P st] for all states [st].  *)

Definition is_wp P c Q :=
  {{P}} c {{Q}} /\
  forall P', {{P'}} c {{Q}} -> (P' ->> P).
```

Agh, this gave me pause. I am just bad at these constructions that do not have any concrete basis. This is quite similar to the things giving me trouble in the Smallstep chapter.

I need to focus. I keep wandering 4chan.

10:30am. Let me finally go through `HoareAsLogic` again.

Focus me, focus.

10:55am.

```
  - eapply H_Consequence with
     (P' := wp (WHILE b DO c END) Q).
    + apply H_While.
      apply IHc.
      intros st st' H [HP Hb] st'' Hwhile.
      apply HP.
      apply (E_WhileTrue st st' st''); assumption.
    + intros st HP st' H.
      apply (HT st st'); assumption.
    + intros st [H Hb].
      apply H.
      apply E_WhileFalse.
      apply bassn_eval_false.
      assumption.
```

Agh...

Even after doing the review I still have absolutely no idea what the hell this proof is supposed to be.

The whole thing is so strange. I can't twist my perspective at all so it makes sense.

Forget it. Maybe at some point I will figure it out. Let me go to the next chapter...after the break."