"9:05am. I am up. Everything was a chore so I went to bed early yesterday. Really, I did nothing right in the past 1.5 decades of my life. I should have aimed to become a real programmer from the start, yet I only ran away. Now I have the vision of what needs to be done, so there is no need to hesitate or look back anymore.
https://mangadex.org/title/26157/otokujuku-gaiden-kurenai-onnajuku

I'll leave this thing for later. Manga, anime, games, they are just a chore. I need to focus on upgrading my skills. Eventually I am going to be done with the SF book.

The SF book is easy, everything is neatly lined up for me there. It is impossible to move from to intermediate to expert without doing my own proofs from start to finish. The same was with Spiral and regular programming.

I do not know how long it will take me to get there, but today I will make another step towards that.

These are the dark ages that I am living in. I put in all this effort, but have nothing to show for it, nor do I have the algorithm that I need. Deep learning lured me in, but did not deliver on its promise.

If I can attain true expertise in theorem proving and make just one crack in the credit assignment problem then my path will be aligned with the future. At that point I will have become a true programmer and will have everything that I need - the skill at programming, the agents of my desire and the future of further progress. If I can solve the issue of making poker agents in a robust manner, I will be able to scale those efforts all the way to the Singularity.

9:25am. Now that I've done a little rant, how about I start? Let me finally finish that fifth chapter. It was a long time since I felt that intense feeling of dissatisfaction with everything as last night. Since I've been doing this for 4.5 years by now, I've settled into a leisurely pace, but radical work should not be treated like a common job. True obsession should be fiery.

Damn, I hate the whole field of ML so much. Their incompetence is making me spend so much of my valuable time on this and yet despite that they do not even have the basics down. With Spiral I've shown the PL guys how it is done. It is their turn next.

There will be nothing left after this."