"8:35am. A little change in plans. I am going to take an hour or two this morning to read the Pluribus paper. Then I'll close the Github issue.
8:50am. Let me chill for a bit and then I will start.

9:10am. Let me start.

9:30am.

```
The blueprint strategy for Pluribus was computed in 8 days on a 64-core server for a total of 12,400 CPU core hours. It required less than 512 GB of memory. At current cloud computing spot instance rates, this would cost about $144 to produce.
```

```
Pluribus to run during live play on a machine with no more than 128 GB of memory while storing a compressed form of the blueprint strategy in memory.
```

This is still quite a lot. My PC only has 8Gb.

9:35am.

```
On average, Pluribus plays at a rate of 20 s per hand when playing against copies of itself in six-player poker. This is roughly twice as fast as professional humans tend to play.
```

No, it really is not when considering pro online play.

9:40am. The paper itself had pretty much nothing, but let me check out the supplementary material.

https://science.sciencemag.org/content/sci/suppl/2019/07/10/science.aay2400.DC1/aay2400-Brown-SM.pdf

10:05am. On page 17 there is the pseudocode for the algorithm. It does not seem too complicated.

10:30am. Done with the supplement. Let me make the closing comment.

11am.

```
Let me add just a bit more, since now I've actually gone through the paper and the supplement. The Monte Carlo CFR on the page 17 of the [supplement](https://science.sciencemag.org/content/sci/suppl/2019/07/10/science.aay2400.DC1/aay2400-Brown-SM.pdf) is quite similar to regular CFR which I've implemented before so it would not be too hard to recreate, but the paper does a lot of fairly hacky engineering to fit what is essentially a tabular algorithm on a game as huge as 6-max NL Holdem. I do not think I would be able to do those parts nearly as easily and I suspect that a lot of those choices are arbitrary. My instinct is to find a way to abstract those parts.

So this is about the standard for ML in the year 2019 - high engineering and complexity, high hype and low theory and understanding. There will be an [AMA](https://www.reddit.com/r/MachineLearning/comments/ceece3/ama_we_are_noam_brown_and_tuomas_sandholm/) with the authors tomorrow and I am almost tempted to ask whether there are any formal proofs demonstrating the superiority of CFR to standard RL. I am not being entirely sarcastic here; I'll have to do a literature review to figure that out at some point.

That having said, tabular algorithms are not a bad pick compared to NNs right now.

On a toy poker game I once trained a RNN with KFAC. When I increased the sequence length feed to the agent from hand-to-hand to game-to-game, the performance of the RNN dropped off a cliff, but tabular RL actually got better. I found this quite impressive. I tried various tricks to make the RNN work better, but I quickly reached my limits. This impressed upon me just how unreliable deep RL really is.

So all in all, given the present state of affairs regular jobs still have appeal compared to banditry. I am just really unfortunate that science needs to be developed before it can be misused. But Spiral is done and the algorithms are not here, so what choice do I have but to do it?

I'll close this thread here, but feel free to reopen it if you have any more questions.
```

Seems fine.

11:05am. Let me take a break here. Since I haven't watched the first lecture like I said I would yesterday, I'll do that once it is time to resume. Then I will do the second chapter. It is best to take it easy.

I am decently good at proofs now, so I should have some fun with my studies from here on out. I gain nothing from rushing it."