"9:10am. Focus me, stop reading the ML sub.
```
Fixpoint select (x: nat) (l: list nat) : nat * list nat :=
match l with
|  nil => (x, nil)
|  h::t => if x <=? h
               then let (j,l') := select x t in (j, h::l')
               else let (j,l') := select h t in (j, x::l')
end.
```

Yesterday I do not think I really understood what this is doing. It occurred to me during the night that `x` might be intended as the head of the list here. This form is one way of ensuring it operates on at least one element.

Before I can start proving things, I need a firm grasp on how this function operates. I played around with it a little yesterday, but let me do it some more now.

```
Compute (select 4 [3;2;1]).
```
```
     = (1, [4; 3; 2])
```

I do not get it. Is that 1 in the output supposed to be the tail of the list? Is the first argument supposed to be the input and the first output the tail?

```
Compute (select 1 [3;2;3]).
```
```
     = (1, [3; 2; 3])
     : nat * list nat
```

Yeah, I am observing this and seeing the code for it, but I am not sure what it is supposed to be doing.

```
Fixpoint selsort l :=
match l with
| i::r => let (j,r') := select i r
               in j :: selsort r'
| nil => nil
```

9:25am. I am getting lost in my daydreams now rather than focusing on the code. Going to bed late will affect me adversely. It is never good to play too much like I did last night, but it is best to treasure such moments of immersion. So I won't regret it.

9:30am. Knowing what particular piece of code does is less than half of what is needed to actually understand it. For complete understanding it is at least as important to know what it should do. Programming is not evolution - it has intention behind it. And that intention is what cuts through the fog.

In the above fragment I can see that `j` is always placed at the head, meaning it has to be the minimum. And `r'` is passed into the `selsort` recursively, meaning it has no assumption behind it.

So `selsort` even though it has nothing to do with `select` in terms of its implementation, is actually more important to understanding `select` than it itself.

Now that I understand this much, let me take a look at `select` again.

```
Fixpoint select (x: nat) (l: list nat) : nat * list nat :=
match l with
|  nil => (x, nil)
|  h::t => if x <=? h
               then let (j,l') := select x t in (j, h::l')
               else let (j,l') := select h t in (j, x::l')
end.
```

Yeah, yesterday I did not really understand the intention behind the `(j, x::l')` fragment, but there is no special reasoning behind `x :: l'`. The main thing is to just propagate the minimum and not drop any of the elements from the original list. The principles of progress and preservation need to be maintained. Science often rhymes.

Great. Now I understand selection sort 100%. Let me just do a little backtracking.

```
Fixpoint insert (i:nat) (l: list nat) :=
  match l with
  | nil => i::nil
  | h::t => if i <=? h then i::h::t else h :: insert i t
 end.

Fixpoint sort (l: list nat) : list nat :=
  match l with
  | nil => nil
  | h::t => insert h (sort t)
end.
```

Insertion sort is quite similar to selection sort isn't it?

Would I be able to think up all this stuff on my own at the level I am currently at? At some point in order to attain true expertise, I am going to have to attempt my own projects. The understanding one gets from that cannot even be compared to doing prepared exercises, but since I am a baby right now, I guess it is fine if I continue this for the time being. The handholding period will not go on for much longer.

9:45am. Let me give the exercise a try again."