"3:40pm. https://arduino.stackexchange.com/questions/816/c-vs-the-arduino-language
So it is just C++ with a bad IDE.

```
The Arduino language is C++, but it is very different from most C++ varieties. The Arduino language has a lot of abstraction built in, especially in the hardware interfaces, which makes it very simple to use. If you have a background in Java, C and C++ should be very similar.

The main differences between Arduino and C++ are in the memory storage. Usually a modern computer has more than 2GB of RAM, while the Arduino Uno has 2kB (1 million times less). The Arduino also uses 8-bit instructions in stead of the 32 bit ones a computer uses. This will mainly affect the amount of information you can store in a variable.
```

Spiral should be good for this sort of thing. Still, if we are talking 2kb, that really limits what the robot should be capable of in a big way.

3:55pm. https://en.wikipedia.org/wiki/List_of_Arduino_boards_and_compatible_systems

My estimate of how much resources I would have at my disposal is off by orders of magnitude it seems. This kind of resource availability really pushes the equation towards custom logic. When he mentioned sensory integration I immediately started thinking about ML solutions.

4pm. https://www.youtube.com/watch?v=nL34zDTPkcs
You can learn Arduino in 15 minutes.

Let me watch some videos on this.

4:20pm. Seems simple enough. So the big innovation with Arduino from a programming perspective is that one can use C? High level programming that Spiral allows would really help with both memory and speed. Hypothetically, if I am going to be doing this for a while then it makes sense to put some effort into adapting Spiral for this.

https://www.youtube.com/watch?v=yohYrKCexvM
3 Creative ideas with Arduino

Let me watch this as well.

4:40pm. https://www.youtube.com/watch?v=OZTRV0D62As
Arduino Projects - Lekcija 1: Uvod u programiranje

This one is 40m. I guess I'll top the day off with this. Maybe it is good that I am fatigued as it gives me time to deal with this. So far it has been moderately interesting. I would not mind doing this for money, but I am too busy with PL work to play with this.

5pm. https://youtu.be/OZTRV0D62As?t=726
"Da li želimo našom dobrom voljom da ga pomažemo Arduino projektu razvoju. Ja kažem da ne želimo pomoći pošto smo mi iz Srbije i nemamo puno para. Jedva imamo za hleb tako da ćemo zamo bezobrazno skinuti i nećemo dati nikakvu donaciju."

This part made me laugh. He is the translation.

"Do we want out of our good will to help the development of the Arduino project. I say that we do not want to help since we are from Serbia and we do not have much money. We barely have for bread so we will rudely download and won't give any donation."

Truly excellent delivery on the joke.

5:15pm. https://youtu.be/OZTRV0D62As?t=1663

I just realized that the bottom of the screen shows the memory usage. So 2kb main memory and 32kb of instruction memory. Lot of inlining potential then.

I know that I really should not be focusing on this since it is so unlikely that even if I ended up working on this that I will have the time to port Spiral, but thinking of how I could use Spiral to break new ground gives me energy.

I suppose years down the road I will know enough type theory to create a dependently typed language and will be able to fuse dependent types and staging, but right now I still really like Spiral for what it is.

5:20pm. Hmm...724 bytes used for just two nearly empty functions. I also remember reading that Arduino uses 1 byte instructions. So I wonder why the long setup?

5:40pm. Done for today.

It occurred to me just because Arduino has low capacity does not mean you can't do AI stuff with it. You could set it as a receiver and control it wirelessly from a PC. That would give you all the capacity you need to do RL and such.

Yesterday's session plus overdone gaming session plus trouble falling sleep really wrecked me today that I've only thought of this now. So it makes no sense to port Spiral to this and all the points about not wanting to do robotics are still valid.

5:45pm. I'll do a better job tomorrow. I will at least finish this chapter and a chunk of the next one.

Today, I might not have done much, but at least I figured out which of my assumptions was broken and significantly improved my understanding of induction. This chimes with my philosophy that if I can't do much that I should at least do what I can."